+++ spec/requests/admin_qualification_report_upload_spec.rb
@@ -28,15 +28,18 @@ RSpec.describe "Admin qualification report upload" do
     end
 
     context "when the data in CSV matches the data in the claim" do
-      it "creates qualification task for the claim" do
+      it "creates qualification and identity_confirmation tasks for the claim" do
         expect {
           post admin_qualification_report_uploads_path, params: {file: file}
-        }.to change { claim.tasks.count }.by(1)
+        }.to change { claim.tasks.count }.by(2)
 
         qualification_task = claim.tasks.find_by(name: "qualifications")
         expect(qualification_task.created_by).to eql(@signed_in_user)
 
-        expect(flash[:notice]).to eql("DQT report uploaded successfully. Automatically completed 1 task for 1 checked claim.")
+        id_confirmation_task = claim.tasks.find_by(name: "identity_confirmation")
+        expect(id_confirmation_task.created_by).to eql(@signed_in_user)
+
+        expect(flash[:notice]).to eql("DQT report uploaded successfully. Automatically completed 2 tasks for 1 checked claim.")
         expect(response).to redirect_to(admin_claims_path)
       end
     end